Add GCE DinD based CI test job for IPv6
This change adds a CI test job for testing Kubernetes IPv6-only functionality.
This CI test uses deployment scripts from the Mirantis/kubeadm-dind-cluster
github repo to set up a containerized (Docker-in-Docker) Kubernetes
cluster that is running on a GCE instance. (A containerized environment
is needed because GCE does not support IPv6 connectivity between GCE
instances.

Note that when the kubeadm-dind-cluster scripts are used to spin up
a Kubernetes cluster in IPv6 mode, it will create the following
container on the GCE instance:
- bind9 container that performs DNS64 functionality
- tayga container that performs NAT64 functionality
- kube-master node container
- kube-node-1 node container
- kube-node-2 node container